Added option for using multiprocessing instead of htcondor
